refactor: 각 인풋 컴포넌트화,css 분리
각 인풋 컴포넌트화 작업
- props 객체로 모아서 한번에 전달
css 분리
- 각 컴포넌트에 해당하는 css 분리
- 기존의 css는 공용 css만 남김
