+++ src/pages/addItem/TagInput.module.css
@@ -1,0 +1,34 @@
+.btnClose {
+  background-image: url("../../assets//btn-close.svg");
+  width: 22px;
+  height: 24px;
+  margin-left: 8px;
+}
+
+.containerTags {
+  width: 100%;
+  display: flex;
+  gap: 12px;
+  flex-wrap: wrap;
+}
+
+.tag {
+  min-width: 100px;
+  height: 48px;
+  border-radius: 26px;
+  background-color: #f9fafb;
+  border: 1px solid #f9fafb;
+  display: flex;
+  justify-content: space-around;
+  align-items: center;
+  padding: 0 8px 0 8px;
+}
+
+.tag span {
+  text-wrap: wrap;
+  font-family: Pretendard;
+  font-size: 16px;
+  font-weight: 400;
+  line-height: 24px;
+  color: #1f2937;
+}
